// src/app/models/usuario.model.ts
export class UsuarioModel{
    id: string;
    nombre: string;
    apellido1: string;
    apellido2: string;
    telefono: string;
    correo: string;
    cp: number;
    colonia: string;
    estado: string;
    ciudad: string;
    calle: string;
    numero: string;
    curp: string;
    usuario: string;
    contrasena: string;
    rol: string;
    cedprod: string;
    especialidad: string;
    foto: string;
    constructor(){
    }
}